requirements.txt
appnope==0.1.3
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
Django==4.2.3
django-filter==23.3
django-utils-six==2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==1.2.0
flex==6.14.1
idna==3.4
inflection==0.5.1
ipykernel==6.23.1
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
jsonpointer==2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
rfc3987==1.3.8
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
strict-rfc3339==0.7
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
validate-email==1.3
wcwidth==0.2.6
wrapt==1.15.0
